requirements.txt
numpy
pandas
matplotlib
torch
Pillow
tqdm
wandb